requirements.txt
channels==4.1.0
dj_database_url==2.2.0
Django==3.2.4
Django==4.2.13
django_environ==0.11.2
djangorestframework_simplejwt==5.3.1
drf_yasg==1.21.7